requirements.txt
flask~=3.0.3
python-dotenv~=1.0.1
azure-cognitiveservices-speech
pytest
pymongo